+++ Source/BugsnagClient.m
@@ -69,8 +69,8 @@ struct bugsnag_data_t {
     // Contains properties in the Bugsnag payload overridden by the user before
     // it was sent
     char *userOverridesJSON;
-    // User onError handler
-    void (*onError)(const BSG_KSCrashReportWriter *writer);
+    // User onCrash handler
+    void (*onCrash)(const BSG_KSCrashReportWriter *writer);
 };
 
 static struct bugsnag_data_t bsg_g_bugsnag_data;
@@ -118,7 +118,7 @@ void BSSerializeDataCrashHandler(const BSG_KSCrashReportWriter *writer, int type
         }
         if (crashSentinelPath != NULL) {
             // Create a file to indicate that the crash has been handled by
-            // the library. This exists in case the subsequent `onError` handler
+            // the library. This exists in case the subsequent `onCrash` handler
             // crashes or otherwise corrupts the crash report file.
             int fd = open(crashSentinelPath, O_RDWR | O_CREAT, 0644);
             if (fd > -1) {
@@ -127,8 +127,8 @@ void BSSerializeDataCrashHandler(const BSG_KSCrashReportWriter *writer, int type
         }
     }
 
-    if (bsg_g_bugsnag_data.onError) {
-        bsg_g_bugsnag_data.onError(writer);
+    if (bsg_g_bugsnag_data.onCrash) {
+        bsg_g_bugsnag_data.onCrash(writer);
     }
 }
 
@@ -265,7 +265,7 @@ void BSGWriteSessionCrashData(BugsnagSession *session) {
         self.crashSentry = [BugsnagCrashSentry new];
         self.errorReportApiClient = [[BugsnagErrorReportApiClient alloc] initWithConfig:configuration
                                                                               queueName:@"Error API queue"];
-        bsg_g_bugsnag_data.onError = (void (*)(const BSG_KSCrashReportWriter *))self.configuration.onError;
+        bsg_g_bugsnag_data.onCrash = (void (*)(const BSG_KSCrashReportWriter *))self.configuration.onCrashHandler;
 
         static dispatch_once_t once_t;
         dispatch_once(&once_t, ^{
@@ -359,6 +359,7 @@ NSString *const BSGBreadcrumbLoadedMessage = @"Bugsnag loaded";
                       onCrash:&BSSerializeDataCrashHandler];
     [self computeDidCrashLastLaunch];
     [self setupConnectivityListener];
+    [self updateAutomaticBreadcrumbDetectionSettings];
 
     NSNotificationCenter *center = [NSNotificationCenter defaultCenter];
     [self watchLifecycleEvents:center];
@@ -428,10 +429,12 @@ NSString *const BSGBreadcrumbLoadedMessage = @"Bugsnag loaded";
 
     [self.sessionTracker startNewSessionIfAutoCaptureEnabled];
 
-    [self addBreadcrumbWithBlock:^(BugsnagBreadcrumb *_Nonnull breadcrumb) {
-        breadcrumb.type = BSGBreadcrumbTypeState;
-        breadcrumb.message = BSGBreadcrumbLoadedMessage;
-    }];
+    if ([self.configuration shouldRecordBreadcrumbType:BSGBreadcrumbTypeState]) {
+        [self addBreadcrumbWithBlock:^(BugsnagBreadcrumb *_Nonnull breadcrumb) {
+            breadcrumb.type = BSGBreadcrumbTypeState;
+            breadcrumb.message = BSGBreadcrumbLoadedMessage;
+        }];
+    }
 
     // notification not received in time on initial startup, so trigger manually
     [self willEnterForeground:self];
@@ -535,6 +538,11 @@ NSString *const BSGBreadcrumbLoadedMessage = @"Bugsnag loaded";
     [self.errorReportApiClient flushPendingData];
 }
 
+/**
+ * Monitor the Bugsnag endpoint to detect changes in connectivity,
+ * flush pending events when (re)connected and report connectivity
+ * changes as breadcrumbs, if configured to do so.
+ */
 - (void)setupConnectivityListener {
     NSURL *url = self.configuration.notifyURL;
 
@@ -549,14 +557,15 @@ NSString *const BSGBreadcrumbLoadedMessage = @"Bugsnag loaded";
         if (connected)
             [strongSelf flushPendingReports];
 
-        [strongSelf addBreadcrumbWithBlock:^(BugsnagBreadcrumb *crumb) {
-            crumb.message = @"Connectivity change";
-            crumb.type = BSGBreadcrumbTypeState;
-            crumb.metadata  = @{ @"type"  :  connectionType };
-        }];
+        if ([[self configuration] shouldRecordBreadcrumbType:BSGBreadcrumbTypeState]) {
+            [strongSelf addBreadcrumbWithBlock:^(BugsnagBreadcrumb *crumb) {
+                crumb.message = @"Connectivity change";
+                crumb.type = BSGBreadcrumbTypeState;
+                crumb.metadata  = @{ @"type"  :  connectionType };
+            }];
+        }
     }];
 }
-
 
 - (void)notifyError:(NSError *)error
               block:(void (^)(BugsnagEvent *))block {
@@ -585,17 +594,6 @@ NSString *const BSGBreadcrumbLoadedMessage = @"Bugsnag loaded";
                      block(report);
                  }
                }];
-}
-
-- (void)notifyException:(NSException *)exception
-             atSeverity:(BSGSeverity)severity
-                  block:(void (^)(BugsnagEvent *))block {
-
-    BugsnagHandledState *state = [BugsnagHandledState
-        handledStateWithSeverityReason:UserSpecifiedSeverity
-                              severity:severity
-                             attrValue:nil];
-    [self notify:exception handledState:state block:block];
 }
 
 - (void)notifyException:(NSException *)exception
@@ -666,8 +664,9 @@ NSString *const BSGBreadcrumbLoadedMessage = @"Bugsnag loaded";
 }
 
 - (void)notify:(NSException *)exception
-    handledState:(BugsnagHandledState *_Nonnull)handledState
-           block:(void (^)(BugsnagEvent *))block {
+  handledState:(BugsnagHandledState *_Nonnull)handledState
+         block:(void (^)(BugsnagEvent *))block
+{
     NSString *exceptionName = exception.name ?: NSStringFromClass([exception class]);
     NSString *message = exception.reason;
     if (handledState.unhandled) {
@@ -712,16 +711,20 @@ NSString *const BSGBreadcrumbLoadedMessage = @"Bugsnag loaded";
                                    config:[self.configuration.config toDictionary]
                              discardDepth:depth];
 
-    [self addBreadcrumbWithBlock:^(BugsnagBreadcrumb *_Nonnull crumb) {
-      crumb.type = BSGBreadcrumbTypeError;
-      crumb.message = reportName;
-      crumb.metadata = @{
-          BSGKeyMessage : reportMessage,
-          BSGKeySeverity : BSGFormatSeverity(report.severity)
-      };
-    }];
+    if ([self.configuration shouldRecordBreadcrumbType:BSGBreadcrumbTypeError]) {
+        [self addBreadcrumbWithBlock:^(BugsnagBreadcrumb *_Nonnull crumb) {
+          crumb.type = BSGBreadcrumbTypeError;
+          crumb.message = reportName;
+          crumb.metadata = @{
+              BSGKeyMessage : reportMessage,
+              BSGKeySeverity : BSGFormatSeverity(report.severity)
+          };
+        }];
+    }
     [self flushPendingReports];
 }
+
+// MARK: - Breadcrumbs
 
 - (void)addBreadcrumbWithBlock:
     (void (^_Nonnull)(BugsnagBreadcrumb *_Nonnull))block {
@@ -762,18 +765,21 @@ NSString *const BSGBreadcrumbLoadedMessage = @"Bugsnag loaded";
     }
 }
 
+/**
+ * Update the device status in response to a battery change notification
+ *
+ * @param notification The change notification
+ */
 #if TARGET_OS_TV
 #elif TARGET_IPHONE_SIMULATOR || TARGET_OS_IPHONE
-- (void)batteryChanged:(NSNotification *)notif {
-    NSNumber *batteryLevel =
-            @([UIDevice currentDevice].batteryLevel);
-    NSNumber *charging =
-            @([UIDevice currentDevice].batteryState ==
-                    UIDeviceBatteryStateCharging);
+- (void)batteryChanged:(NSNotification *)notification {
+    NSNumber *batteryLevel = @([UIDevice currentDevice].batteryLevel);
+    NSNumber *charging = @([UIDevice currentDevice].batteryState == UIDeviceBatteryStateCharging);
 
     [[self state] addAttribute:BSGKeyBatteryLevel
                      withValue:batteryLevel
                  toTabWithName:BSGKeyDeviceState];
+    
     [[self state] addAttribute:BSGKeyCharging
                      withValue:charging
                  toTabWithName:BSGKeyDeviceState];
@@ -821,6 +827,15 @@ NSString *const BSGBreadcrumbLoadedMessage = @"Bugsnag loaded";
         }
     }
 
+    if ([[self configuration] shouldRecordBreadcrumbType:BSGBreadcrumbTypeState]) {
+        [self addBreadcrumbWithBlock:^(BugsnagBreadcrumb *_Nonnull breadcrumb) {
+            breadcrumb.type = BSGBreadcrumbTypeState;
+            breadcrumb.message = orientationNotifName;
+            breadcrumb.metadata = @{BSGKeyOrientation : orientation};
+        }];
+    }
+
+    // Update the Device orientation
     [[self state] addAttribute:BSGKeyOrientation
                      withValue:orientation
                  toTabWithName:BSGKeyDeviceState];
@@ -831,6 +846,10 @@ NSString *const BSGBreadcrumbLoadedMessage = @"Bugsnag loaded";
                      withValue:[[Bugsnag payloadDateFormatter]
                                    stringFromDate:[NSDate date]]
                  toTabWithName:BSGKeyDeviceState];
+     
+    if ([[self configuration] shouldRecordBreadcrumbType:BSGBreadcrumbTypeState]) {
+        [self sendBreadcrumbForNotification:notif];
+    }
 }
 #endif
 
@@ -848,6 +867,53 @@ NSString *const BSGBreadcrumbLoadedMessage = @"Bugsnag loaded";
         [self.oomWatchdog disable];
     }
 }
+
+/**
+ * Configure event listeners (i.e. observers) for enabled automatic breadcrumbs.
+ */
+- (void)updateAutomaticBreadcrumbDetectionSettings {
+   // State events
+    if ([[self configuration] shouldRecordBreadcrumbType:BSGBreadcrumbTypeState]) {
+        // Generic state events
+        for (NSString *name in [self automaticBreadcrumbStateEvents]) {
+            [self startListeningForStateChangeNotification:name];
+        }
+        
+        // NSMenu events (Mac only)
+        for (NSString *name in [self automaticBreadcrumbMenuItemEvents]) {
+            [[NSNotificationCenter defaultCenter]
+                addObserver:self
+                   selector:@selector(sendBreadcrumbForMenuItemNotification:)
+                       name:name
+                     object:nil];
+        }
+    }
+    
+    // Navigation events
+    if ([[self configuration] shouldRecordBreadcrumbType:BSGBreadcrumbTypeNavigation]) {
+        // UI/NSTableView events
+        for (NSString *name in [self automaticBreadcrumbTableItemEvents]) {
+            [[NSNotificationCenter defaultCenter]
+                addObserver:self
+                   selector:@selector(sendBreadcrumbForTableViewNotification:)
+                       name:name
+                     object:nil];
+        }
+    }
+
+    // User events
+    if ([[self configuration] shouldRecordBreadcrumbType:BSGBreadcrumbTypeUser]) {
+        // UITextField/NSControl events (text editing)
+        for (NSString *name in [self automaticBreadcrumbControlEvents]) {
+            [[NSNotificationCenter defaultCenter]
+                addObserver:self
+                   selector:@selector(sendBreadcrumbForControlNotification:)
+                       name:name
+                     object:nil];
+        }
+    }
+}
+
 - (NSArray<NSString *> *)automaticBreadcrumbStateEvents {
 #if TARGET_OS_TV
     return @[
@@ -931,7 +997,12 @@ NSString *const BSGBreadcrumbLoadedMessage = @"Bugsnag loaded";
 #endif
 }
 
-- (void)crumbleNotification:(NSString *)notificationName {
+/**
+ * Configure a generic state change breadcrumb listener
+ *
+ * @param notificationName The name of the notification.
+ */
+- (void)startListeningForStateChangeNotification:(NSString *)notificationName {
     [[NSNotificationCenter defaultCenter]
         addObserver:self
            selector:@selector(sendBreadcrumbForNotification:)
@@ -950,13 +1021,18 @@ NSString *const BSGBreadcrumbLoadedMessage = @"Bugsnag loaded";
     }];
 }
 
-- (void)sendBreadcrumbForTableViewNotification:(NSNotification *)note {
+/**
+ * Leave a navigation breadcrumb whenever a tableView selection changes
+ *
+ * @param notification The UI/NSTableViewSelectionDidChangeNotification
+ */
+- (void)sendBreadcrumbForTableViewNotification:(NSNotification *)notification {
 #if TARGET_IPHONE_SIMULATOR || TARGET_OS_IPHONE || TARGET_OS_TV
-    UITableView *tableView = [note object];
+    UITableView *tableView = [notification object];
     NSIndexPath *indexPath = [tableView indexPathForSelectedRow];
     [self addBreadcrumbWithBlock:^(BugsnagBreadcrumb *_Nonnull breadcrumb) {
       breadcrumb.type = BSGBreadcrumbTypeNavigation;
-      breadcrumb.message = BSGBreadcrumbNameForNotificationName(note.name);
+      breadcrumb.message = BSGBreadcrumbNameForNotificationName(notification.name);
       if (indexPath) {
           breadcrumb.metadata =
               @{ @"row" : @(indexPath.row),
@@ -964,10 +1040,10 @@ NSString *const BSGBreadcrumbLoadedMessage = @"Bugsnag loaded";
       }
     }];
 #elif TARGET_OS_MAC
-    NSTableView *tableView = [note object];
+    NSTableView *tableView = [notification object];
     [self addBreadcrumbWithBlock:^(BugsnagBreadcrumb *_Nonnull breadcrumb) {
       breadcrumb.type = BSGBreadcrumbTypeNavigation;
-      breadcrumb.message = BSGBreadcrumbNameForNotificationName(note.name);
+      breadcrumb.message = BSGBreadcrumbNameForNotificationName(notification.name);
       if (tableView) {
           breadcrumb.metadata = @{
               @"selectedRow" : @(tableView.selectedRow),
@@ -978,15 +1054,20 @@ NSString *const BSGBreadcrumbLoadedMessage = @"Bugsnag loaded";
 #endif
 }
 
-- (void)sendBreadcrumbForMenuItemNotification:(NSNotification *)notif {
+/**
+* Leave a state breadcrumb whenever a tableView selection changes
+*
+* @param notification The UI/NSTableViewSelectionDidChangeNotification
+*/
+- (void)sendBreadcrumbForMenuItemNotification:(NSNotification *)notification {
 #if TARGET_OS_TV
 #elif TARGET_IPHONE_SIMULATOR || TARGET_OS_IPHONE
 #elif TARGET_OS_MAC
-    NSMenuItem *menuItem = [[notif userInfo] valueForKey:@"MenuItem"];
+    NSMenuItem *menuItem = [[notification userInfo] valueForKey:@"MenuItem"];
     if ([menuItem isKindOfClass:[NSMenuItem class]]) {
         [self addBreadcrumbWithBlock:^(BugsnagBreadcrumb *_Nonnull breadcrumb) {
           breadcrumb.type = BSGBreadcrumbTypeState;
-          breadcrumb.message = BSGBreadcrumbNameForNotificationName(notif.name);
+          breadcrumb.message = BSGBreadcrumbNameForNotificationName(notification.name);
           if (menuItem.title.length > 0)
               breadcrumb.metadata = @{BSGKeyAction : menuItem.title};
         }];